+++ R/globals.R
@@ -3,6 +3,7 @@
 # source: https://community.rstudio.com/t/how-to-solve-no-visible-binding-
 #         for-global-variable-note/28887
 my_globals <- c(
+  ".data",
   "anulRedRate", "maturity", "minRedQty",
   "minRedVal", "min_qtd", "min_value", "mtrtyDt",
   "name", "nm", "price", "untrRedVal", "annual_ret"